AccountsTest updated and not required test cases moved to new test files
