+++ packages/web3-eth/tests/src/signers/TransactionSignerTest.js
@@ -1,0 +1,17 @@
+// Mocks
+jest.mock('');
+
+/**
+ * TransactionSignerTest test
+ */
+describe('TransactionSignerTestTest', () => {
+    let TransactionSignerTest;
+
+    beforeEach(() => {
+        TransactionSignerTest = new TransactionSignerTest();
+    });
+
+    it('constructor check', () => {
+
+    });
+});
